Do not propagate context's statement_of_account value to avoid performance issues.
Rule's domain_get() uses a cache system where it only takes into account the
user and context's datetime. This made the system perform poorly when the first
time the user accessed move lines was the statements of account screen because
the context included the 'statement_of_account' value and the the module
override search() to include order by clauses.

Those clauses are unnecessary on Rule's and caused the performance issues.

Now we ensure that when create, write, read and search operations are called,
the super is called without the context value so that when the rule system
calls account.move.line search(), the order by will not be added.